pyproject.toml
[project]
name = "chatdify"
version = "0.0.1.beta.1"
description = "A Python connector for integrating Chatwoot with Dify AI"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "celery>=5.4.0",
    "redis>=5.2.1",
    "asyncpg>=0.30.0",
    "sentry-sdk[fastapi,celery,sqlalchemy,httpx]>=2.24.0",
]

[dependency-groups]
dev = [
    "ruff",
    "pre-commit",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "pytest-cov>=4.0.0",
    "pytest-dotenv>=0.5.2",
    "sqlalchemy2-stubs>=0.0.2a38",
    "mypy>=1.0.0",
    "sqlmodel>=0.0.22", # Temporary for existing tests - to be migrated in separate task
    "aiosqlite>=0.21.0",
]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = ["B008", "W293"]  # fastapi dependency injection needs B008, W293 for trailing whitespace

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"
env_files = [".env"]
